python_consumer/requirements.txt
requests
pika 
openai-whisper
pydub
difflib
python-magic